// generated/src/app/api-models/pmb-subscriber-snapshot.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSubscriberSnapshot {
  seqSubsId: number;
  seqSubsUpdate: number;
  effectiveDate: Date;
  planCode: string;
  familySize: number;
  spouseFlag: string;
  seqGroupId: number;
  otherStatusFlag: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  termDate: Date;
  subscriberDob: Date;
  subscriberGender: string;
  salary: number;
  medicareStatus: string;
  subscDept: string;
  subscLocation: string;
  billEffectiveFrom: Date;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  rateType: string;
}